pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "modl"
version = "0.1.0"
description = "Package, run, and serve AI/ML models locally"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "fastapi>=0.111",
    "uvicorn>=0.30",
]

[project.scripts]
modl = "modl.cli.app:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["modl*"]
